make html2pdf more reliable

+++ plugins/plugins-available/reports2/script/html2pdf.sh
@@ -39,6 +39,18 @@ for x in seq 10; do
     [ -e $XAUTHORITY ] && break;
 done
 
+# wait x-lock
+for x in seq 5; do
+    sleep 1;
+    [ -e /tmp/.X${DISP}-lock ] && break;
+done
+
+if [ ! -e /tmp/.X${DISP}-lock -o ! -e $XAUTHORITY ]; then
+    echo "xvfb failed to start"
+    cat $TMPLOG
+    exit 1
+fi
+
 rm -f $OUTPUT
 DISPLAY=:$DISP $WKHTMLTOPDF \
         --use-xserver \
@@ -62,4 +74,4 @@ if [ -e "$OUTPUT" -a $UID == 0 ]; then
 fi
 
 kill $xpid >/dev/null 2>&1
-rm -f $TMPLOG $XAUTHORITY
+rm -f $TMPLOG $XAUTHORITY /tmp/.X${DISP}-lock
